+++ src/main/java/com/example/demo/entity/Spot.java
@@ -20,7 +20,7 @@ public class Spot {
 	private String name;
 	private String category;
 
-	private Integer rating; // 1〜5
+	private Double rating;
 	@Column(length = 1000)
 	private String comment;
 
@@ -56,11 +56,11 @@ public class Spot {
 		this.category = category;
 	}
 
-	public Integer getRating() {
+	public Double getRating() {
 		return rating;
 	}
 
-	public void setRating(Integer rating) {
+	public void setRating(Double rating) {
 		this.rating = rating;
 	}
